LayoutTests/imported/w3c/web-platform-tests/dom/nodes/MutationObserver-cross-realm-callback-report-exception-expected.txt

CONSOLE MESSAGE: Error: PASS


PASS MutationObserver reports the exception from its callback in the callback's global object
